// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-education-details-model.ts
export class EmployeeEducationDetailsModel{
    employeeEducationDetailId: string;
    employeeId: string;
    educationLevelId: string;
    institute: string;
    majorSpecialization: string;
    gpaOrScore: number;
    startDate: Date;
    endDate: Date;
    isArchived: boolean;
    operationsPerformedBy: string;

    firstName: string;
    surName: string;
    userName: string;
    email: string;
    educationLevel: string;
    timeStamp: any;
    totalCount: number;

}